Refactor: refactor removal of cbse
