+++ Collision/src/main/resources/META-INF/entityprocessor.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="dk.sdu.mmmi.cbse.osgicollision">
-    <implementation class="dk.sdu.mmmi.cbse.osgicollision.Collider"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="dk.sdu.mmmi.collision">
+    <implementation class="dk.sdu.mmmi.collision.Collider"/>
     <service>
-        <provide interface="dk.sdu.mmmi.cbse.common.services.IPostEntityProcessingService"/>
+        <provide interface="dk.sdu.mmmi.common.services.IPostEntityProcessingService"/>
     </service>
 </scr:component>
