+++ Common/osgi.bnd
@@ -1,1 +1,1 @@
-Export-Package: dk.sdu.mmmi.cbse.common.*
+Export-Package: dk.sdu.mmmi.common.*
