--- Common/src/main/java/dk/sdu/mmmi/cbse/common/data/entityparts/EntityPart.java
@@ -1,8 +1,0 @@
-package dk.sdu.mmmi.cbse.common.data.entityparts;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.GameData;
-
-public interface EntityPart {
-    void process(GameData gameData, Entity entity);
-}
