+++ Common/src/main/java/dk/sdu/mmmi/common/data/Entity.java
@@ -1,6 +1,6 @@
-package dk.sdu.mmmi.cbse.common.data;
+package dk.sdu.mmmi.common.data;
 
-import dk.sdu.mmmi.cbse.common.data.entityparts.EntityPart;
+import dk.sdu.mmmi.common.data.entityparts.EntityPart;
 import java.io.Serializable;
 import java.util.Map;
 import java.util.UUID;
