+++ Common/src/main/java/dk/sdu/mmmi/common/data/GameData.java
@@ -1,6 +1,6 @@
-package dk.sdu.mmmi.cbse.common.data;
+package dk.sdu.mmmi.common.data;
 
-import dk.sdu.mmmi.cbse.common.events.Event;
+import dk.sdu.mmmi.common.events.Event;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
