+++ Common/src/main/java/dk/sdu/mmmi/common/data/World.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.cbse.common.data;
+package dk.sdu.mmmi.common.data;
 
 import java.util.ArrayList;
 import java.util.Collection;
