+++ Common/src/main/java/dk/sdu/mmmi/common/data/entityparts/EntityPart.java
@@ -1,0 +1,8 @@
+package dk.sdu.mmmi.common.data.entityparts;
+
+import dk.sdu.mmmi.common.data.Entity;
+import dk.sdu.mmmi.common.data.GameData;
+
+public interface EntityPart {
+    void process(GameData gameData, Entity entity);
+}
