+++ Common/src/main/java/dk/sdu/mmmi/common/data/entityparts/PositionPart.java
@@ -1,7 +1,7 @@
-package dk.sdu.mmmi.cbse.common.data.entityparts;
+package dk.sdu.mmmi.common.data.entityparts;
 
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.GameData;
+import dk.sdu.mmmi.common.data.Entity;
+import dk.sdu.mmmi.common.data.GameData;
 
 public class PositionPart implements EntityPart {
