+++ Core/src/main/java/dk/sdu/mmmi/core/managers/GameInputProcessor.java
@@ -1,9 +1,9 @@
-package dk.sdu.mmmi.cbse.core.managers;
+package dk.sdu.mmmi.core.managers;
 
 import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.InputAdapter;
-import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.GameKeys;
+import dk.sdu.mmmi.common.data.GameData;
+import dk.sdu.mmmi.common.data.GameKeys;
 
 public class GameInputProcessor extends InputAdapter {
