+++ Core/src/main/resources/META-INF/core.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="dk.sdu.mmmi.cbse.Game">
-    <implementation class="dk.sdu.mmmi.cbse.Game"/>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="dk.sdu.mmmi.Game">
+    <implementation class="dk.sdu.mmmi.Game"/>
 
     <reference bind="addPostEntityProcessingService" cardinality="0..n"
-               interface="dk.sdu.mmmi.cbse.common.services.IPostEntityProcessingService"
+               interface="dk.sdu.mmmi.common.services.IPostEntityProcessingService"
                name="IPostEntityProcessingService" policy="dynamic" unbind="removePostEntityProcessingService"/>
 
     <reference bind="addEntityProcessingService" cardinality="0..n"
-               interface="dk.sdu.mmmi.cbse.common.services.IEntityProcessingService"
+               interface="dk.sdu.mmmi.common.services.IEntityProcessingService"
                name="IEntityProcessingService" policy="dynamic" unbind="removeEntityProcessingService"/>
 
     <reference bind="addGamePluginService" cardinality="0..n"
-               interface="dk.sdu.mmmi.cbse.common.services.IGamePluginService"
+               interface="dk.sdu.mmmi.common.services.IGamePluginService"
                name="IGamePluginService" policy="dynamic" unbind="removeGamePluginService"/>
 </scr:component>
